+++ inst/@mp/ctranspose.m
@@ -1,4 +1,4 @@
 function result = ctranspose (varargin)
   error ("mplapack:NotImplemented", ...
-         "mp conjugate transpose is not implemented before matrix storage is finalized in M07");
+         "mp conjugate transpose is not implemented in M07");
 endfunction